@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.4 <4.0.0'
+  sdk: ">=3.9.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,15 +34,16 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.6
+  cupertino_icons: ^1.0.8
   flutter_sound: ^9.4.6
-  permission_handler: ^11.3.1
-  path_provider: ^2.1.3
+  permission_handler: ^12.0.1
+  path_provider: ^2.1.5
   loading_indicator: ^3.1.1
   flutter_keyboard_visibility: ^6.0.0
-  rxdart: ^0.27.7
-  lottie: ^3.1.0
-  just_audio: ^0.9.37
+  rxdart: ^0.28.0
+  lottie: ^3.3.2
+  just_audio: ^0.10.5
+  logger: ^2.6.2
 
 dev_dependencies:
   flutter_test:
@@ -53,7 +54,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^3.0.0
+  flutter_lints: ^6.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
